+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the package will be documented here.
 ## [4.26.1] - 2025-11-08
 
 - **Fixed** timezone conversion issues by updating `TIME_ZONE_IDS` constant: **normalized** all negative signs (replaced `'−'` with `'-'`) from UTC offset.
+- **Redesigned** *time zone determiners* with updated *timezone constants*. **Added** *3 overload signatures* for `timeZone(utc?)` method.
  -->
 
 ## [4.26.1] - 2025-11-08